Move ensure_instance_included to djangae.db.consistency
Also make sure that the included instance is only returned if it matches the filters. Update the documentation.
